+++ crates/rpc/rpc/src/lib.rs
@@ -6,6 +6,7 @@ mod rpc;
 mod web3;
 
 pub use debug::DebugApi;
+pub use eth::EthApi;
 pub use net::NetApi;
 pub use rpc::RPCApi;
 pub use web3::Web3Api;
